requirements.txt
asgiref==3.11.1
black==25.11.0
click==8.1.8
Django==4.2.30
django-rest-framework==0.1.0
djangorestframework==3.16.1
flake8==7.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.4.0
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.4.1
sqlparse==0.5.5
tomli==2.4.1
typing_extensions==4.15.0
